pyproject.toml
[tool.poetry]
name = "specnut"
version = "0.1.0"
description = "CLI tool for optimizing specification digests to reduce token consumption"
authors = ["SpecNut Team <team@specnut.dev>"]
readme = "README.md"
homepage = "https://github.com/specnut/specnut"
repository = "https://github.com/specnut/specnut"
keywords = ["cli", "tokens", "optimization", "specifications", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.21.0"}
tiktoken = "^0.8.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
ruff = "^0.0.287"

[tool.poetry.scripts]
specnut = "specnut.cli.commands:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=specnut --cov-report=html --cov-report=term"
